+++ src/cjs/VEllipseSector.js
@@ -10,6 +10,7 @@
  * @modified 2025-04-01 Adapting a the `toCubicBezier` calculation to match an underlying change in the vertAt and tangentAt calculation of ellipses (was required to hamonize both methods with circles).
  * @modified 2025-04-02 Adding `VEllipseSector.containsAngle` method.
  * @modified 2025-04-02 Adding `VEllipseSector.lineIntersections` and `VEllipseSector.lineIntersectionTangents` and implementing `Intersectable`.
+ * @modified 2025-04-07 Adding value wrapping (0 to TWO_PI) to the `VEllipseSector.containsAngle` method.
  * @version  1.2.0
  */
 Object.defineProperty(exports, "__esModule", { value: true });
@@ -64,14 +65,22 @@ var VEllipseSector = /** @class */ (function () {
      * @return {boolean} True if (and only if) this sector contains the given angle.
      */
     VEllipseSector.prototype.containsAngle = function (angle) {
-        // angle -= this.ellipse.rotation;
-        angle = geomutils_1.geomutils.wrapMinMax(angle - this.ellipse.rotation, 0, Math.PI * 2);
-        if (this.startAngle <= this.endAngle) {
-            return angle >= this.startAngle && angle < this.endAngle;
+        angle = geomutils_1.geomutils.mapAngleTo2PI(angle); // wrapMinMax(angle, 0, Math.PI * 2);
+        var sAngle = geomutils_1.geomutils.mapAngleTo2PI(this.startAngle);
+        var eAngle = geomutils_1.geomutils.mapAngleTo2PI(this.endAngle);
+        // TODO: cleanup
+        // if (this.startAngle <= this.endAngle) {
+        //   return angle >= this.startAngle && angle < this.endAngle;
+        // } else {
+        //   // startAngle > endAngle
+        //   return angle >= this.startAngle || angle < this.endAngle;
+        // }
+        if (sAngle <= eAngle) {
+            return angle >= sAngle && angle < eAngle;
         }
         else {
             // startAngle > endAngle
-            return angle >= this.startAngle || angle < this.endAngle;
+            return angle >= sAngle || angle < eAngle;
         }
     };
     //--- BEGIN --- Implement interface `Intersectable`
@@ -95,7 +104,7 @@ var VEllipseSector = /** @class */ (function () {
         return ellipseIntersections.filter(function (intersectionPoint) {
             tmpLine.b.set(intersectionPoint);
             var lineAngle = tmpLine.angle();
-            return _this.containsAngle(lineAngle);
+            return _this.containsAngle(lineAngle - _this.ellipse.rotation);
         });
     };
     /**