+++ src/cjs/utils/helpers/VEllipseSectorHelper.d.ts
@@ -4,6 +4,7 @@
  * @author   Ikaros Kappler
  * @date     2021-03-24
  * @modified 2025-04-02 Adding `VEllipseSectorHelper.drawHandleLines`.
+ * @modified 2025-05-07 Modifying the calculation of `startAngle` and `endAngle` from the rotation control point: wrapping result into [0,TWO_PI).
  * @version  1.1.0
  */
 import { VEllipseSector } from "../../VEllipseSector";